+++ CHANGELOG.md
@@ -1,3 +1,20 @@
+2015.08.22 - [v1.3.0](https://github.com/mivaecommerce/readytheme-suivant/wiki/v1.3.0-Updates)
+---
+*	Updates and additions to support the new features released in Miva 9.4: http://www.miva.com/mm9.4_template_changes.html
+*	Added new pages to "screenCheck" variable in HEAD tag.
+*	Added support for forgot password on WLGN page in scripts.js.
+*	Updated mobile footer function in scripts.js for better iOS user experience and removed screen toggling.
+*	Miscellaneous clean-up to CSS and JS files.
+*	Upgraded css.php to output the concatenated and minified CSS files in a more user-friendly way.
+*	Removed footer navigation toggle on screen resize.
+*	Fixed the following issues:
+	*	[Issue #5](https://github.com/mivaecommerce/readytheme-suivant/issues/5)
+	*	[Issue #6](https://github.com/mivaecommerce/readytheme-suivant/issues/6)
+	*	[Issue #7](https://github.com/mivaecommerce/readytheme-suivant/issues/7)
+	*	[Issue #8](https://github.com/mivaecommerce/readytheme-suivant/issues/8)
+	*	[Issue #9](https://github.com/mivaecommerce/readytheme-suivant/issues/9)
+	*	[Issue #10](https://github.com/mivaecommerce/readytheme-suivant/issues/10)
+
 2015.04.15 - [v1.2.0](https://github.com/mivaecommerce/readytheme-suivant/wiki/v1.2.0-Updates)
 ---
 *	Updates and additions to support the new features released in Miva 9.3: http://www.miva.com/mm9.3_template_changes.html http://extranet.miva.com/forums/showthread.php?115591-Miva-Merchant-9-0003-Is-Now-Available&p=430880#post430880